Conduct Pseudo-bulked Differential Expression Within Cell Types

This function conducts our standard workflow for pseudo-bulked differential gene expression testing within cell types using DESeq2. For a more modular and interactive workflow, see the individual create_pseudobulk(), create_metadata(), create_DESeq(), DESeq(), and results_DESeq() functions. If running this analysis for the first time, we'd recommend running each step interactively so that you can inspect the model fit quality.

within_celltype_DEG(
  seurat_obj,
  cell_type_var,
  grouping_vars,
  counts_filter = 5,
  Design,
  Contrast,
  ...
)

Arguments

seurat_obj

Seurat object

cell_type_var

Chr name of metadata variable containing cell type annotation

grouping_vars

Chr vector of metadata variables that define cell grouping strategy

counts_filter

Int minimum count to retain gene

Design

Formula for model to fit

Contrast

Chr vector of contrast in the format c("variable", "level2", "level1")

...

Additional arguments to pass to DESeq2

Value

Data frame of DESeq results

Examples

within_celltype_DEG(
    seurat_obj = Neurons,
    cell_type_var = "CellType",
    grouping_vars = c("Sex", "GEM"),
    Design = ~Treatment,
    Contrast = c("Treatment", "Mor", "Sal")
)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 't': object 'Neurons' not found